Setup automatic gradle plugin versioning
Summary:
This diff makes the versioning of the Gradle plugin automatic. During gradle build, we're now generating a `version.kt` file that contains a constant pointing to `VERSION_NAME` from `gradle.properties` file.
So when we'll be releasing new versions of the DCG, we'll only have to update the `VERSION_NAME` in `gradle.properties` file and we won't have to update the version in the gradle plugin source file manually.

Reviewed By: rybalkinsd

Differential Revision: D44023734

fbshipit-source-id: 4b992091208dbacce149ad404e5fbf1c91896924