Initial publishing setup for :compiler:* projects.
Summary:
This sets up the publishign to Maven Local for
```
    ":compiler:cli",
    ":compiler:common",
    ":compiler:k1",
    ":compiler:k2",
```

The other libraries are KMP libraries are require a different setup. Similarly
the Gradle plugins will require a different setup as well.

I'll follow-up on them. This is just the setup of a precompiled script plugin
we can use to easily publish any module.

Differential Revision: D44144462

fbshipit-source-id: d5d7a6c470885538f3eec98d4600e851c26c5839