Initial publishing setup for :compiler:* projects. (#5)
Summary:
Pull Request resolved: https://github.com/facebookincubator/dataclassgenerate/pull/5

This sets up the publishign to Maven Local for
```
    ":compiler:cli",
    ":compiler:common",
    ":compiler:k1",
    ":compiler:k2",
```

The other libraries are KMP libraries are require a different setup. Similarly
the Gradle plugins will require a different setup as well.

I'll follow-up on them. This is just the setup of a precompiled script plugin
we can use to easily publish any module.

Reviewed By: zielinskimz

Differential Revision: D44144462

fbshipit-source-id: 1c473b70e95768686fd1096df44792115943a240