Open source DCG tests
Summary:
This diff moves DCG tests from javatests directory to DCG directory and puts them in a directory structure expected by Gradle. javatests directory now contains only 2 BUCK files which reference the tests from the main DCG directory, it's been done this way in order to not break Sandcastle DCG test jobs.

Also the K2 tests have been temporarily disabled because there were issues with running them via Gradle. In order to fix it we would have to bump Kotlin Compile Testing library to 1.5.0 which requires Kotlin 1.8 which is not compatible with 1.7.21. Migration to Kotlin 1.8 at Meta is already in progress: D42585524 so once it's done, we'll be able to re-enable K2 tests.

Reviewed By: adicatana

Differential Revision: D44203730

fbshipit-source-id: 35e3edb5ad8c1afcdd0a7d635fd03998d246e864