DoNotStripAny on DataClassSuper
Summary:
## Context
The no arg constructor on the `DataClassSuper` class is inlined into subclasses and later removed.
This is causing the Deserialization error on IG4A reported on this [post](https://fb.workplace.com/groups/redex.feedback/posts/8137723819630664).

## The Change
This is a temp fix that addresses this particular data class deserialization issue.
We should follow up on the underlying issue that avoid inlining ctor from super class into a Serializable subclass.

Differential Revision: D62903858

fbshipit-source-id: 5a507440fbb032b5809c30ca14b1d045a2fd963c