Adapter and custom buck target for IDE integration of DCG compiler plugin
Summary:
For live edit / code hot swap in Android Studio, we compile classes locally in the IDE and then diff against the buck compiled class to check for unsupported changes (e.g. changing constructor, changing super type). This fails on classes annotated with `DataClassGenerate` since when an FOA class is compiled with buck, the data class generate compiled plugin is run which adds `DataClassSuper` as a supertype of the class.

To ensure the class supertype matches when compiled in the IDE vs when compiled with buck, we need to ensure we run the data class generate compiler plugin when compiling in the IDE as well.

I added support for that in the next diff (D77935185). However, while doing the integration I found that it crashes with a `NoSuchMethodError`  for `fun defineClass` in `DataClassGenerateBuilder`. I looked into it and turns out it's because the IDE and the DGC plugin use different package names for `PsiElement`.

The IDE uses `com.intellij.psi.PsiElement` (from //third-party/java/intellij:plugin-sdk) while the DGC plugin uses ` org.jetbrains.kotlin.com.intellij.psi.PsiElement` (from //third-party/kotlin:kotlin-compiler).

In this diff I'm adding a separate target for the IDE and typealiases for PsiElement to handle this.

Differential Revision: D77935184

fbshipit-source-id: cabf57a3456083b8b4620b1b5c5500659d9193a7